Don't force takeScreenshot to take a size and don't force output to jpeg.
Remove the requirement of a size from takeScreenshot since ffmpeg will
just use the size of the video in the case that it isn't passed.  Also,
don't force the output to be a jpeg.  FFmpeg will figure out the best
option it can from the file extension of the output file.  Plus, use
the newer _spawnProcess method instead of directly calling exec so that
streams can be passed (I'm not 100% sure this works right with streams,
but I was testing a weird case with mongodb and gridfs streams).